refactor(deps-core,deps-swift,deps-github-actions): extract shared GitHub tags-API client
deps-swift and deps-github-actions each reimplemented the same GitHub
tags-API client: owner/repo validation, auth-header/token setup, tags
pagination, and page parsing were duplicated near-verbatim between the
two registries with no shared abstraction.

Extract the common logic into deps_core::github (GithubTagsClient,
validate_owner_repo, is_valid_github_identity, paginate_tags,
parse_tags_page, fetch_tags_page). Both registries now build on the
shared client while keeping their ecosystem-specific extensions
(deps-github-actions' RateLimitGate, in-flight coalescing, and
TagIndex).

Behavior-preserving except for one intentional improvement: a GitHub
Actions tags page with a single entry missing/malformed commit.sha no
longer aborts the whole page's parse, it now skips just that entry.

Resolves #472
