+++ crates/deps-swift/src/lib.rs
@@ -24,26 +24,13 @@ pub use parser::parse_package_swift;
 pub use registry::SwiftRegistry;
 pub use types::{SwiftDependency, SwiftPackage, SwiftParseResult, SwiftVersion};
 
-/// Whether `name` matches the `owner/repo` GitHub identifier shape this crate accepts:
-/// `[a-zA-Z0-9._-]+/[a-zA-Z0-9._-]+`, with neither segment being exactly `.`/`..` (see
-/// [`deps_core::is_dot_segment`]).
+/// Whether `name` matches the `owner/repo` GitHub identifier shape this crate accepts.
 ///
 /// Shared by `registry::validate_owner_repo` (a credential-bearing fetch-URL gate) and
 /// `formatter::is_valid_owner_repo` (a display-URL gate), so the two predicates cannot
-/// drift out of sync on what counts as a valid identity — `registry`'s dot-segment
-/// rejection (#357) would otherwise have applied to the fetch path only, leaving
-/// `formatter`'s hover link still able to render `https://github.com/apple/..`.
+/// drift out of sync on what counts as a valid identity. Delegates to
+/// [`deps_core::github::is_valid_github_identity`], shared with `deps-github-actions`
+/// (#472).
 pub(crate) fn is_valid_github_identity(name: &str) -> bool {
-    static RE: std::sync::OnceLock<regex::Regex> = std::sync::OnceLock::new();
-    let re = RE.get_or_init(|| {
-        regex::Regex::new(r"^[a-zA-Z0-9._-]+/[a-zA-Z0-9._-]+$").expect("hardcoded regex is valid")
-    });
-    match name.split_once('/') {
-        Some((owner, repo)) => {
-            re.is_match(name)
-                && !deps_core::is_dot_segment(owner)
-                && !deps_core::is_dot_segment(repo)
-        }
-        None => false,
-    }
+    deps_core::github::is_valid_github_identity(name)
 }
